pyproject.toml
[tool.poetry]
name = "python-kevo"
version = "0.1.0"
description = "Python SDK for Kevo key-value store"
authors = ["Jeremy Tregunna <jeremy@tregunna.ca>"]
readme = "README.md"
packages = [{include = "kevo"}]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
protobuf = "^4.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
pylint = "^2.17.5"
grpc-stubs = "^1.53.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pylint.messages_control]
disable = "C0111,R0903,C0103"
